backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
Pillow
httpx
typer>=0.12,<1.0
aiofiles
cryptography
premailer
playwright
